Move Static* classes to a separate module.
That doesn't export any `Closure` related things.
